+++ packages/nx/src/command-line/run/run.ts
@@ -5,19 +5,8 @@ import {
 } from '../../utils/params';
 import { printHelp } from '../../utils/print-help';
 import { NxJsonConfiguration } from '../../config/nx-json';
-import { readJsonFile } from '../../utils/fileutils';
-import { buildTargetFromScript, PackageJson } from '../../utils/package-json';
-import { join, relative } from 'path';
-import { existsSync } from 'fs';
-import {
-  loadNxPlugins,
-  mergePluginTargetsWithNxTargets,
-} from '../../utils/nx-plugin';
-import {
-  ProjectConfiguration,
-  TargetConfiguration,
-  ProjectsConfigurations,
-} from '../../config/workspace-json-project-json';
+import { relative } from 'path';
+import { ProjectsConfigurations } from '../../config/workspace-json-project-json';
 import { Executor, ExecutorContext } from '../../config/misc-interfaces';
 import { TaskGraph } from '../../config/task-graph';
 import { serializeOverridesIntoCommandLine } from '../../utils/serialize-overrides-into-command-line';
@@ -90,25 +79,6 @@ async function iteratorToProcessStatusCode(
   }
 }
 
-function createImplicitTargetConfig(
-  root: string,
-  proj: ProjectConfiguration,
-  targetName: string
-): TargetConfiguration | null {
-  const packageJsonPath = join(root, proj.root, 'package.json');
-  if (!existsSync(packageJsonPath)) {
-    return null;
-  }
-  const { scripts, nx } = readJsonFile<PackageJson>(packageJsonPath);
-  if (
-    !(targetName in (scripts || {})) ||
-    !(nx.includedScripts && nx.includedScripts.includes(targetName))
-  ) {
-    return null;
-  }
-  return buildTargetFromScript(targetName, nx);
-}
-
 async function parseExecutorAndTarget(
   { project, target, configuration }: Target,
   root: string,
@@ -116,14 +86,7 @@ async function parseExecutorAndTarget(
   nxJsonConfiguration: NxJsonConfiguration
 ) {
   const proj = projectsConfigurations.projects[project];
-  const targetConfig =
-    proj.targets?.[target] ||
-    createImplicitTargetConfig(root, proj, target) ||
-    mergePluginTargetsWithNxTargets(
-      proj.root,
-      proj.targets,
-      await loadNxPlugins(nxJsonConfiguration.plugins, [root], root)
-    )[target];
+  const targetConfig = proj.targets?.[target];
 
   if (!targetConfig) {
     throw new Error(`Cannot find target '${target}' for project '${project}'`);
