+++ packages/nx/src/utils/globs.ts
@@ -1,6 +1,4 @@
-export const combineGlobPatterns = (patterns: string[]) =>
-  patterns.length > 1
-    ? '{' + patterns.join(',') + '}'
-    : patterns.length === 1
-    ? patterns[0]
-    : '';
+export function combineGlobPatterns(...patterns: (string | string[])[]) {
+  const p = patterns.flat();
+  return p.length > 1 ? '{' + p.join(',') + '}' : p.length === 1 ? p[0] : '';
+}
